requirements.txt
absl-py==2.1.0
aiohttp==3.9.1
aioquic==1.0.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.9.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.3
attrs==22.2.0
bcrypt==4.1.3
beautifulsoup4==4.13.4
blinker==1.6.2
boto3==1.35.44
botocore==1.35.44
Brotli==1.1.0
bs4==0.0.2
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
chex==0.1.85
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.8
cycler==0.12.1
distlib==0.3.6
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.2.5
django-cors-headers==4.7.0
django-graphql-jwt==0.4.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
dpkt==1.9.8
etils==1.7.0
exceptiongroup==1.1.0
filelock==3.9.0
Flask==3.0.3
Flask-Cors==4.0.0
flatbuffers==24.3.7
flax==0.8.2
fonttools==4.48.1
frozenlist==1.4.1
fsspec==2024.3.0
gast==0.5.4
google-auth==2.28.2
google-auth-oauthlib==1.2.0
google-genai==1.11.0
google-pasta==0.2.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
grpcio==1.62.1
h11==0.14.0
h2==4.1.0
h5py==3.10.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.28.1
hyperframe==6.0.1
idna==3.4
importlib_resources==6.3.1
itsdangerous==2.1.2
jax==0.4.25
jaxlib==0.4.25
Jinja2==3.1.2
jmespath==1.0.1
kaitaistruct==0.10
keras==2.15.0
kiwisolver==1.4.5
ldap3==2.9.1
libclang==16.0.6
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mitmproxy==10.3.1
mitmproxy-windows==0.5.2
mitmproxy_rs==0.5.2
ml-dtypes==0.3.2
MouseInfo==0.1.3
msgpack==1.0.8
multidict==6.0.4
nest-asyncio==1.6.0
ngrok==0.12.1
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.2.1
orbax-checkpoint==0.5.6
outcome==1.2.0
packaging==23.2
pandas==2.2.1
paramiko==3.4.0
passlib==1.7.4
pillow==10.2.0
platformdirs==3.0.0
promise==2.3
protobuf==4.25.3
psutil==6.0.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
publicsuffix2==2.20191221
pyaes==1.6.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.53
pycparser==2.21
pydantic==2.11.3
pydantic_core==2.33.1
pydivert==2.1.0
PyGetWindow==0.0.9
Pygments==2.17.2
PyJWT==2.10.1
pylsqpack==0.3.18
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyOpenSSL==24.1.0
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pyshark==0.6
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.0
python-telegram-bot==13.7
pytweening==1.0.4
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.28.2
requests-file==2.1.0
requests-oauthlib==1.4.0
rich==13.7.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.3
scapy==2.5.0
scipy==1.12.0
SCons==4.7.0
selenium==4.8.2
service-identity==24.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.4.3
Telethon==1.33.1
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.1
tensorflow-estimator==2.15.0
tensorflow-hub==0.16.1
tensorflow-intel==2.15.1
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_decision_forests==1.8.1
tensorflowjs==4.17.0
tensorstore==0.1.56
termcolor==2.4.0
text-unidecode==1.3
tf_keras==2.15.1
tldextract==5.1.2
toolz==0.12.1
tornado==6.4
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2022.7
tzlocal==5.2
urllib3==1.26.14
urwid-mitmproxy==2.1.2.1
virtualenv==20.19.0
webdriver-manager==3.8.5
websockets==15.0.1
Werkzeug==3.0.3
wrapt==1.14.1
wsproto==1.2.0
wurlitzer==3.0.3
yarl==1.9.4
zipp==3.18.1
zstandard==0.22.0
